+++ packages/garden/src/lib/components/FilterSelector/FilterSelector.tsx
@@ -26,7 +26,7 @@ const getFieldSettingsLabelFromKey = <
 
 export function FilterSelector(): JSX.Element | null {
 	const controller = useGardenContext();
-	const { setHorizontalGroupingAccessor, setVerticalGroupingKeys, customViews, data, fieldSettings } = controller;
+	const { setHorizontalGroupingAccessor, setVerticalGroupingKeys, customViews, getData, fieldSettings } = controller;
 
 	const { gardenKey, groupByKeys } = useGroupingKeys();
 
@@ -45,8 +45,8 @@ export function FilterSelector(): JSX.Element | null {
 		() =>
 			fieldSettings && Object.keys(fieldSettings).length
 				? Object.keys(fieldSettings)
-				: Object.keys(data?.value[0] as Record<string, unknown>),
-		[fieldSettings, data.value]
+				: Object.keys(getData()[0] as Record<string, unknown>),
+		[fieldSettings, getData]
 	);
 
 	const filterGroupKey = useCallback(
@@ -56,14 +56,14 @@ export function FilterSelector(): JSX.Element | null {
 
 	const groupingOptions = useMemo(
 		(): string[] =>
-			data.value.length
+			getData().length
 				? allOptions
 						.filter(filterGroupKey)
 						.map((groupKey) => fieldSettings?.[groupKey]?.label || groupKey)
 						.sort()
 				: [],
 
-		[data, fieldSettings, filterGroupKey, allOptions]
+		[getData, fieldSettings, filterGroupKey, allOptions]
 	);
 
 	const handleExistingSelectionChange = useCallback(
@@ -92,7 +92,7 @@ export function FilterSelector(): JSX.Element | null {
 		},
 		[fieldSettings, setGardenKey]
 	);
-	if (!data) return null;
+	if (!getData()) return null;
 
 	return (
 		<StyledSelectRowWrapper>