+++ packages/workspace-fusion-modules/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.18](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-fusion-modules@0.1.17...@equinor/workspace-fusion-modules@0.1.18) (2022-12-09)
+
+**Note:** Version bump only for package @equinor/workspace-fusion-modules
+
+## [0.1.17](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-fusion-modules@0.1.16...@equinor/workspace-fusion-modules@0.1.17) (2022-12-09)
+
+**Note:** Version bump only for package @equinor/workspace-fusion-modules
+
 ## [0.1.16](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-fusion-modules@0.1.15...@equinor/workspace-fusion-modules@0.1.16) (2022-12-07)
 
 **Note:** Version bump only for package @equinor/workspace-fusion-modules